requirements.txt
google-auth-oauthlib 
google-api-python-client
google-auth-httplib2
gspread
python-dotenv
pandas